--- Assets/Scripts/Enemy.cs
@@ -1,23 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Enemy : MonoBehaviour
-{
-    /*public float speed = 3.0f;
-    private Rigidbody2D enemyRb2D;
-    GameObject player;
-    // Start is called before the first frame update
-    void Start()
-    {
-         enemyRb2D = GetComponent<Rigidbody2D>();
-         player = GameObject.Find("Player");
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        Vector3 lookDirection = (player.transform.position - transform.position).normalized;
-        enemyRb2D.AddForce(lookDirection * speed * Time.deltaTime);
-    }*/
-}
