--- Assets/Scripts/Enemy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 19b1c14331d154fd7833bdd87b3dc407
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
