+++ Assets/Scripts/Game/NewBehaviourScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3c7c3f3cc25f1406ea7a2a50be6ac9a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
