+++ Assets/Scripts/Game/VolumeManager.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class VolumeManager : MonoBehaviour
+{
+    [SerializeField] Slider volumeSlider;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+    
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
