+++ Assets/Scripts/Game/VolumeManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 43302c638eaca4f83a02fd95078f1e23
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
