--- Assets/Scripts/PauseManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: aa43b809e11f74fdf8cf70371d9a65e9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
