+++ Assets/Scripts/Player/CameraFollow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ad44201c86edd40458515ed6d109ed10
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
