+++ Assets/Scripts/Player/PlayerController.cs
@@ -1,0 +1,36 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerController : MonoBehaviour
+{
+    public float speed = 10.0f;
+    Rigidbody2D rigidbody2d;
+    Animator animator;
+    float horizontal;
+    float vertical;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+         rigidbody2d = GetComponent<Rigidbody2D>();
+         animator = GetComponent<Animator>();
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        horizontal = Input.GetAxis("Horizontal");
+        vertical = Input.GetAxis("Vertical");
+          Vector2 move = new Vector2(horizontal, vertical);
+    }
+    void FixedUpdate()
+    {
+        Vector2 position = rigidbody2d.position;
+        position.x = position.x + speed * horizontal * Time.deltaTime;
+        position.y = position.y + speed * vertical * Time.deltaTime;
+
+        rigidbody2d.MovePosition(position);
+    }
+
+}
