--- Assets/Scripts/StartButtonScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: aae048a01a21542fe943a3eb56f122d8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
