+++ Library/StateCache/SceneView/8c/8cd7c613bf844de3b80696e27a479d5e.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":true,"pivot":{"x":0.7944561243057251,"y":-5.85006046295166,"z":-85.06881713867188},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":19.745153427124025,"orthographic":true}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":true,"pivot":{"x":-2.3963875770568849,"y":-0.7478984594345093,"z":-95.3590087890625},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":22.70509147644043,"orthographic":true}
